feat(android-webview): 添加全面的 WebView 配置选项
- 扩展 AndroidWebViewOptions，涵盖核心功能、缩放、安全、缓存、布局及交互设置
- 更新 NativeBridgeSettings 以解析并应用来自 Unity 的新配置字段
- 在 TransparentWebViewFragment 中应用配置，包含布局算法和缩放控制
- 在 WebBridge.load 中添加日志以调试配置加载
- 修复 UnityBridge 中日志条目操作码的类型转换
- 在 WebMessageDispatcher 中添加错误处理以捕获反序列化失败
- 更新 HTML viewport meta 标签以优化移动端缩放

+++ Native/Android/NativeBridge/src/main/java/com/chaosstudio/nativebridge/UnityBridge.kt
@@ -26,7 +26,7 @@ class UnityBridge {
             logEntry.put("level", level)
             logEntry.put("message", message)
 
-            document.put("opcode", NativeBridgeSettings.webViewOptions.logEntryOpcode)
+            document.put("opcode", NativeBridgeSettings.webViewOptions.logEntryOpcode.toInt())
             document.put("payload", logEntry)
 
             UnityPlayer.UnitySendMessage(UnityGameObjectName, UnityMethodName, document.toString())
